test(graph): share the per-instance state suite between Graph and BaseGraph
BaseGraph.test.ts and Graph.test.ts each checked that the properties coming from the mixins are not shared between
graph instances, with two diverging copies of the same suite. BaseGraph used hand-written tests that omitted
pageFormat and warningImage and relied on deep inequality, while Graph used a data-driven table that also asserts the
mutation took effect, so that a leak cannot pass unnoticed.

The properties under test are declared on AbstractGraph, so the guarantee is identical for every concrete
implementation. Keeping two copies meant a new property had to be added twice, and in practice it was added only to
one of them.

Extract the table and its assertions into describeNoGlobalStateForMixinProperties(), a suite factory parameterized by
a graph factory, and call it from both files. selectionModel joins the table as its read/mutate/signature shape fits.
The helper is a plain .ts file, so Jest does not collect it as a suite of its own.
