+++ packages/core/__tests__/view/no-global-state-for-mixin-properties.ts
@@ -1,0 +1,127 @@
+/*
+Copyright 2026-present The maxGraph project Contributors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import { describe, expect, test } from '@jest/globals';
+import { type AbstractGraph, Cell, Multiplicity } from '../../src';
+
+// `mixInto` installs mixin members on the AbstractGraph prototype, so a mutable default (object or
+// array) declared in a mixin would be shared by every graph instance: mutating it on one graph would
+// mutate it for all of them, with no compile error to signal it.
+//
+// The properties below are declared directly in AbstractGraph, in the "Variables that should be in the
+// mixins but requiring per-instance initialization" group, precisely to avoid that. These tests pin that
+// guarantee down, so moving one of them into a mixin fails here instead of silently corrupting state.
+//
+// When adding a property to that group, add a case here too.
+type PerInstanceProperty = {
+  name: string;
+  /** Reads the property under test on a graph. */
+  read: (graph: AbstractGraph) => object;
+  /** Alters the value in place, the way user code would. */
+  mutate: (value: never) => void;
+  /** Extracts what the mutation is expected to change, to compare two graphs without deep equality. */
+  signature: (value: never) => unknown;
+};
+
+const perInstanceProperties: PerInstanceProperty[] = [
+  {
+    name: 'alternateEdgeStyle',
+    read: (graph) => graph.alternateEdgeStyle,
+    mutate: (value: { fillColor?: string }) => (value.fillColor = 'red'),
+    signature: (value: { fillColor?: string }) => value.fillColor,
+  },
+  {
+    name: 'cells',
+    read: (graph) => graph.cells,
+    mutate: (value: Cell[]) => value.push(new Cell()),
+    signature: (value: Cell[]) => value.length,
+  },
+  {
+    name: 'mouseListeners',
+    read: (graph) => graph.mouseListeners,
+    mutate: (value: object[]) =>
+      value.push({ mouseDown: () => {}, mouseMove: () => {}, mouseUp: () => {} }),
+    signature: (value: object[]) => value.length,
+  },
+  {
+    name: 'multiplicities',
+    read: (graph) => graph.multiplicities,
+    mutate: (value: Multiplicity[]) =>
+      value.push(new Multiplicity(true, 'type', null, null, 0, 1, null, null, null)),
+    signature: (value: Multiplicity[]) => value.length,
+  },
+  {
+    name: 'options',
+    read: (graph) => graph.options,
+    mutate: (value: { foldingEnabled: boolean }) => (value.foldingEnabled = false),
+    signature: (value: { foldingEnabled: boolean }) => value.foldingEnabled,
+  },
+  {
+    name: 'pageFormat',
+    read: (graph) => graph.pageFormat,
+    mutate: (value: { width: number }) => (value.width = 42),
+    signature: (value: { width: number }) => value.width,
+  },
+  {
+    name: 'warningImage',
+    read: (graph) => graph.warningImage,
+    mutate: (value: { src: string }) => (value.src = 'mutated.png'),
+    signature: (value: { src: string }) => value.src,
+  },
+  // Not part of the AbstractGraph group above, but shares the same failure mode: even though
+  // SelectionMixin declares `selectionModel: null` on the prototype, the null value is harmless because
+  // initializeCollaborators assigns a new GraphSelectionModel. The assignment creates a per-instance
+  // property that shadows the prototype null, and GraphSelectionModel's constructor allocates its own
+  // `cells = []` array.
+  {
+    name: 'selectionModel',
+    read: (graph) => graph.getSelectionModel(),
+    mutate: (value: { cells: Cell[] }) => value.cells.push(new Cell()),
+    signature: (value: { cells: Cell[] }) => value.cells.length,
+  },
+];
+
+/**
+ * Registers the suite checking that no mutable state is shared between two graphs created by
+ * `createGraph`.
+ *
+ * Called from the test files of each concrete graph implementation, as the shared state, if any, comes
+ * from the mixins applied to the `AbstractGraph` prototype and is therefore observable from all of them.
+ */
+export const describeNoGlobalStateForMixinProperties = (
+  createGraph: () => AbstractGraph
+): void => {
+  describe('Expect no global state for properties coming from mixins', () => {
+    test.each(perInstanceProperties)('$name', ({ read, mutate, signature }) => {
+      const graph1 = createGraph();
+      const graph2 = createGraph();
+
+      const value1 = read(graph1) as never;
+      const value2 = read(graph2) as never;
+
+      expect(value1).not.toBe(value2);
+      expect(signature(value1)).toStrictEqual(signature(value2));
+
+      const graph2SignatureBeforeMutation = signature(value2);
+      mutate(value1);
+
+      // the mutation is effective, so the assertion below cannot pass by accident
+      expect(signature(value1)).not.toStrictEqual(graph2SignatureBeforeMutation);
+      // and it did not leak to the other graph
+      expect(signature(value2)).toStrictEqual(graph2SignatureBeforeMutation);
+    });
+  });
+};
